updates to v0.3.3

+++ bin/augustus_parallel.py
@@ -8,7 +8,7 @@ class MyFormatter(argparse.ArgumentDefaultsHelpFormatter):
     def __init__(self,prog):
         super(MyFormatter,self).__init__(prog,max_help_position=48)
 parser=argparse.ArgumentParser(prog='augustus_parallel.py', usage="%(prog)s [options] -i genome.fasta -s botrytis_cinera -o new_genome",
-    description='''Script that does it all...''',
+    description='''Script runs augustus in parallel to use multiple processors''',
     epilog="""Written by Jon Palmer (2016) nextgenusfs@gmail.com""",
     formatter_class = MyFormatter)
 parser.add_argument('-i','--input', required=True, help='Genome in FASTA format')
@@ -16,6 +16,7 @@ parser.add_argument('-o','--out', required=True, help='Basename of output files'
 parser.add_argument('-s','--species', required=True, help='Augustus species name')
 parser.add_argument('--hints', help='Hints file (PE)')
 parser.add_argument('--cpus', default=2, type=int, help='Number of CPUs to run')
+parser.add_argument('--debug', action='store_true', help='Keep intermediate files')
 args=parser.parse_args()
 
 #check for augustus installation
@@ -43,14 +44,25 @@ def countGFFgenes(input):
 
 def runAugustus(Input):
     FNULL = open(os.devnull, 'w')
+    if '_part' in Input:
+        chr = Input.split('_part')[0]
+    else:
+        chr = Input
     species='--species='+args.species
-    hints_input = '--hintsfile='+os.path.join(tmpdir, Input+'.hints.gff')
-    aug_out = os.path.join(tmpdir, Input+'.augustus.gff3')   
+    hints_input = '--hintsfile='+os.path.join(tmpdir, chr+'.hints.gff')
+    aug_out = os.path.join(tmpdir, Input+'.augustus.gff3')
+    core_cmd = ['augustus', species, '--gff3=on', '--UTR=off', '--stopCodonExcludedFromCDS=False', os.path.join(tmpdir, chr+'.fa')]
+    if args.hints:
+        core_cmd.insert(2, extrinsic)
+        core_cmd.insert(3, hints_input)
+    if Input in ranges:
+        start = ranges.get(Input)[0]
+        end = ranges.get(Input)[1]
+        core_cmd.insert(2, '--predictionStart='+str(start))
+        core_cmd.insert(3, '--predictionEnd='+str(end))
+
     with open(aug_out, 'w') as output:
-        if args.hints:
-            subprocess.call(['augustus', species, hints_input, extrinsic, '--gff3=on', '--UTR=off', '--stopCodonExcludedFromCDS=False', os.path.join(tmpdir, Input+'.fa')], stdout = output, stderr= FNULL)
-        else:
-            subprocess.call(['augustus', species, '--gff3=on', '--UTR=off', '--stopCodonExcludedFromCDS=False', os.path.join(tmpdir, Input+'.fa')], stdout = output, stderr = FNULL)
+        subprocess.call(core_cmd, stdout = output)
 
 
 #first step is to split input fasta file into individual files in tmp folder
@@ -58,30 +70,56 @@ print("Splitting contigs and hints files")
 tmpdir = 'augustus_tmp_'+str(os.getpid())
 os.makedirs(tmpdir)
 scaffolds = []
+global ranges
+ranges = {}
 with open(args.input, 'rU') as InputFasta:
     for record in SeqIO.parse(InputFasta, 'fasta'):
-        name = str(record.id)
-        scaffolds.append(name)
-        outputfile = os.path.join(tmpdir, name+'.fa')
-        with open(outputfile, 'w') as output:
-            SeqIO.write(record, output, 'fasta')
+        contiglength = len(record.seq)
+        if contiglength > 500000: #split large contigs
+            num_parts = contiglength / 500000 + 1
+            chunks = contiglength / num_parts
+            for i in range(0,num_parts):
+                name = str(record.id)+'_part'+str(i+1)
+                scaffolds.append(name)
+                outputfile = os.path.join(tmpdir, str(record.id)+'.fa')
+                if i == 0: #this is first record
+                    start = 1
+                    end = chunks + 10000
+                else:
+                    start = end - 10000
+                    end = start + chunks + 10000
+                if end > contiglength:
+                    end = contiglength
+                if not name in ranges:
+                    ranges[name] = (start, end)
+                with open(outputfile, 'w') as output:
+                    SeqIO.write(record, output, 'fasta')     
+        else:
+            name = str(record.id)
+            scaffolds.append(name)
+            outputfile = os.path.join(tmpdir, name+'.fa')
+            with open(outputfile, 'w') as output:
+                SeqIO.write(record, output, 'fasta')
 
 #if hints file passed, split it up by scaffold
 if args.hints:
     for i in scaffolds:
-        with open(os.path.join(tmpdir, i+'.hints.gff'), 'w') as output:
-            with open(args.hints, 'rU') as hintsfile:
-                for line in hintsfile:
-                    cols = line.split('\t')
-                    if cols[0] == i:
-                        output.write(line)
+        if '_part' in i:
+            i = i.split('_part')[0]
+        if not os.path.isfile(os.path.join(tmpdir, i+'.hints.gff')):
+            with open(os.path.join(tmpdir, i+'.hints.gff'), 'w') as output:
+                with open(args.hints, 'rU') as hintsfile:
+                    for line in hintsfile:
+                        cols = line.split('\t')
+                        if cols[0] == i:
+                            output.write(line)
 
 #now loop through each scaffold running augustus
 if args.cpus > len(scaffolds):
     num = len(scaffolds)
 else:
     num = args.cpus
-print("Running augustus on %i scaffolds, using %i CPUs" % (len(scaffolds), num))
+print("Running augustus on %i chunks, using %i CPUs" % (len(scaffolds), num))
 p = multiprocessing.Pool(num)
 rs = p.map_async(runAugustus, scaffolds)
 p.close()
@@ -101,5 +139,6 @@ join_script = os.path.join(AUGUSTUS_BASE, 'scripts', 'join_aug_pred.pl')
 with open(args.out, 'w') as finalout:
     with open(os.path.join(tmpdir, 'augustus_all.gff3'), 'rU') as input:
         subprocess.call([join_script],stdin = input, stdout = finalout)
-shutil.rmtree(tmpdir)
+if not args.debug:
+    shutil.rmtree(tmpdir)
 print("Found %i total gene models" % countGFFgenes(args.out))
